cleanup step defs
